test: split agent service coverage
